fix(security+models): CRITICAL/HIGH タスク群を一括修正
## CRITICAL
- TASK-007: StaffTypeTrimmer / BillingStatusPending / PaymentMethodElectronicMoney ENUM定数を追加
- TASK-008: Billing.Subtotal/TaxTotal/TotalAmount を int → int64 に修正（overflow防止）
- TASK-015: LIFF_MOCK が release モードでも有効になるバグを修正（gin.Mode() チェック追加）
- TASK-016: SetPermissionGroupRules に clinicID 所有権チェックを追加（横断テナント書き換え防止）

## HIGH
- TASK-006: appointment_repository の JOIN 条件に reservation_types.deleted_at IS NULL を追加
- TASK-011: billing_item / examination_repository の JOIN 条件に deleted_at IS NULL を追加
- TASK-004: reservation_type_group Update を buildReservationTypeGroupUpdateFields パターンに統一
- TASK-005: reservation_type_group Delete に slog.InfoContext を追加
- TASK-023: reservation_type_group Create の slog bare string キーを slog.Attr に修正・Update 後ログ追加
- TASK-021: useMutation の onError で handleApiError を呼び出すよう修正（4 mutation）

+++ backend/internal/handler/accounting_request.go
@@ -8,9 +8,9 @@ type createAccountingRequest struct {
 	HospitalizationID *uint64    `json:"hospitalization_id"`
 	OwnerID           *uint64    `json:"owner_id"`
 	PetID             *uint64    `json:"pet_id"`
-	Subtotal          int        `json:"subtotal"      binding:"min=0"`
-	TaxTotal          int        `json:"tax_total"     binding:"min=0"`
-	TotalAmount       int        `json:"total_amount"  binding:"min=0"`
+	Subtotal          int64      `json:"subtotal"      binding:"min=0"`
+	TaxTotal          int64      `json:"tax_total"     binding:"min=0"`
+	TotalAmount       int64      `json:"total_amount"  binding:"min=0"`
 	HasInsurance      bool       `json:"has_insurance"`
 	Status            string     `json:"status"`
 	ScheduledDate     time.Time  `json:"scheduled_date" binding:"required"`
@@ -26,9 +26,9 @@ type updateAccountingRequest struct {
 	HospitalizationID *uint64    `json:"hospitalization_id"`
 	OwnerID           *uint64    `json:"owner_id"`
 	PetID             *uint64    `json:"pet_id"`
-	Subtotal          *int       `json:"subtotal"`
-	TaxTotal          *int       `json:"tax_total"`
-	TotalAmount       *int       `json:"total_amount"`
+	Subtotal          *int64     `json:"subtotal"`
+	TaxTotal          *int64     `json:"tax_total"`
+	TotalAmount       *int64     `json:"total_amount"`
 	HasInsurance      *bool      `json:"has_insurance"`
 	Status            *string    `json:"status"`
 	ScheduledDate     *time.Time `json:"scheduled_date"`
